==> code/1_Angiosperm_island_diversity.Rmd <==
---
title: "Angiosperm island diversity"
output: html_document
  code_folding: show
date: "2023-09-21"
---

# Overview

Aim: Put insular Asteraceae diversity in context with the island diversity of other flowering plant (angiosperm) families.

This script will address three main questions:

1.  What are the most diverse flowering plant (angiosperm) families on islands? All islands and oceanic islands.

2.  How does insular Asteraceae diversity compare to other species rich angiosperm families?

3.  Is Asteraceae over or under presented in insular floras? ie How does the number of island species compare to the null expectations?

Main steps (Outline):

1.  Calculate number of island species (native and endemic) for each family

    1.  Diversity on island all islands

    2.  Diversity on oceanic islands

    3.  Angiosperm diversity globally

2.  Binomial tests

3.  Null models and island simulations

4.  Results: Visualize insular diversity and null expectation for angiosperms

## Preparation

Load libraries:

```{r}
library(tidyverse)
library(readxl)
library(writexl)
library(VGAM)
library(ggdist)
library(cowplot)
library(patchwork)
library(svglite)
```

Load data:

Load the GIFT3 checklists of all angiosperm species native to islands.

```{r}
Angiosperms_islands_checklists <- read_excel("../data/data_clean/Angiosperms_islands_checklists.xlsx", guess_max = 10000)
```

Load the island environmental data which will be used to identify continental/oceanic islands. `geology_original` is an internal version of island geology and island age generated by GIFT team during earlier versions of GIFT. `geology_Aster_clean` is a cleaned version of the geology database for all the GIFT3 entities that have native Asteraceae.

```{r}
geology_original <- read.csv("../data/data_raw/geology.csv")

geology_Aster_clean <- read_excel("../data/data_clean/Asteraceae_island_env.xlsx") %>%
  select(c("entity_ID", "archipelago", "longitude", "latitude", "geology", "island_type"))
```

Read in the World Checklist of Vascular Plants (WCVP) database (extracted 10/2022).

```{r}
wcvp <-  read.table("../data/data_raw/wcvp/wcvp_names.csv", sep="|", header=TRUE, quote = "", fill=TRUE, encoding = "UTF-8") 
```

# 1. Calculate number of island species (native and endemic) for each family

## Diversity on all island types (continental, oceanic)

Count the number of unique species per family.

```{r}
native_islands <- Angiosperms_islands_checklists %>%
  group_by(family) %>%
  summarize(native = n_distinct(work_species)) %>%
  arrange(desc(native))

endemic_islands <- Angiosperms_islands_checklists %>%
  filter(endemic_list == 1) %>%
  group_by(family) %>%
  summarize(endemic = n_distinct(work_species)) %>%
  arrange(desc(endemic))
```

Combine into one table

```{r}
diversity_islands_all <- left_join(
  native_islands,
  endemic_islands,
  by = "family"
)
```

## Diversity on oceanic islands

### Island type (oceanic, continental) classification

In order to calculate diversity on islands first we need an accurate classification of island types.

Start by combining the Angiosperm checklists with the cleaned geology data (see `0_data_preparation.Rmd` script).

```{r}
Angiosperms_islands_checklists <- left_join(Angiosperms_islands_checklists, 
                                            geology_Aster_clean,
                                            by = "entity_ID")
```

Next get a list of all the unique entity_IDs that do not have an island_type (oceanic, continental, mixed) assessment.

```{r}
Angiosperm_islands_nogeo <- Angiosperms_islands_checklists %>%
  filter(is.na(island_type)) %>%
  distinct(entity_ID, .keep_all = TRUE) %>%
  select(c("entity_ID", "geo_entity"))
  
```

402 entities have native Angiosperm species but no Asteraceae, and so no cleaned assessment of island type.

Combine the original GIFT geology database to see if there are any more matches.

```{r}
Angiosperm_islands_nogeo <- left_join(
  Angiosperm_islands_nogeo,
  geology_original,
  by = "entity_ID") 

# select on the columns needed now 
Angiosperm_islands_nogeo <- Angiosperm_islands_nogeo %>%
  select(c("entity_ID", "geo_entity", "geology"))
```

Now we need to assess the island type for these 402 entities. Start with a coarse classification of oceanic islands. Leverage GIFT's geology classification (ie atoll, volcanic shelf, etc) in `geology_original.csv` to coarsely classify islands to oceanic or continental.

Identify and classify island type (first coarse pass): Create new `island_type_coarse` column and designate based on `geology` column. Generalizing the geology column to oceanic, continental, (and mixed)

-   oceanic \<=\> atoll, floor, volcanic, floor/volcanic, atoll/volcanic

-   continental \<=\> fragment, shelf

-   other mixed combinations need to be addressed individually:

    -   fragment/volcanic,

    -   shelf/volcanic,

    -   atoll/shelf/volcanic,

    -   atoll/floor/fragment,

    -   fragment/shelf/volcanic,

    -   atoll/floor/fragment/shelf/volcanic

```{r}
Angiosperm_islands_nogeo <- Angiosperm_islands_nogeo %>%
  mutate(
    island_type_coarse = case_when(
      geology == "atoll" ~ "oceanic", # specify when geology cases oceanic
      geology == "floor" ~ "oceanic",
      geology == "volcanic" ~ "oceanic",
      geology == "floor/volcanic" ~ "oceanic",
      geology == "atoll/volcanic" ~ "oceanic",
      is.na(geology) ~ NA_character_, # maintain NAs as true NAs
      geology == "shelf" ~ "continental", # specify when geology cases non oceanic
      geology == "fragment" ~ "continental"
      ))

```

Now we will review each island assessment manually by checking the coarse assessment and adding an assessment for the NAs. Write out for manual review:

```{r}
write_xlsx(Angiosperm_islands_nogeo, "../data/data_raw/Angiosperm_islands_nogeo_dirty.xlsx")
```

Read in the cleaned database of islands with no prior island type assessment (Manual review done by Lizzie Roeble, September 2023).

```{r}
# read in cleaned database and only take entity_ID and island_type_clean 
Angiosperm_islands_nogeo_cleaned <- read_xlsx("../data/data_raw/Angiosperm_islands_nogeo_clean.xlsx") %>%
  select(c("entity_ID", "island_type_clean"))
```

Get the new island_type assessments integrated into the `Angiosperms_islands_checklists`. First join the tables where `island_type_clean` is a seperate column.

```{r}
Angiosperms_islands_checklists <- left_join(
  Angiosperms_islands_checklists,
  Angiosperm_islands_nogeo_cleaned,
  by = "entity_ID")
```

Next, use coalesce() (<https://dplyr.tidyverse.org/reference/coalesce.html>) to replace the NA values in `island_type` with the new assessments from `island_type_clean`.

```{r}
# add new assessments with coalesce
Angiosperms_islands_checklists$island_type <- coalesce(Angiosperms_islands_checklists$island_type, Angiosperms_islands_checklists$island_type_clean) 

# remove the island_type_clean column
Angiosperms_islands_checklists <- Angiosperms_islands_checklists %>%
  select(!island_type_clean)

```

Check that all entity_IDs have a island type classification

```{r}
sum(is.na(Angiosperms_islands_checklists$island_type))
```

### Calculate diversity on oceanic islands

Filter checklists to only those on oceanic islands.

```{r}
Angiosperms_islands_oceanic_checklists <- Angiosperms_islands_checklists %>%
  filter(island_type == "oceanic")
```

Calculate diversity (number of native and endemic species) for each angiosperm family on oceanic islands.

```{r}
native_oceanic <- Angiosperms_islands_oceanic_checklists %>%
  group_by(family) %>%
  summarize(native = n_distinct(work_species)) %>%
  arrange(desc(native))

endemic_oceanic <- Angiosperms_islands_oceanic_checklists %>%
  filter(endemic_list == 1) %>%
  group_by(family) %>%
  summarize(endemic = n_distinct(work_species)) %>%
  arrange(desc(endemic))
```

Combine into one table.

```{r}
diversity_islands_oceanic <- left_join(
  native_oceanic,
  endemic_oceanic,
  by = "family"
)
```

## Angiosperm diversity globally

In order to analyze whether island diversity is more or less than expected given a families overall diversity globally, we need the accepted number of species for each family. Getting estimates for the number of species in a plant family is not trivial. There are several taxonomic databases, each with their own strengths and weaknesses. The most common taxonomic databases include: The Plant List ([http://www.theplantlist.org/),](http://www.theplantlist.org/),) World Flora Online ([https://wfoplantlist.org/plant-list),](https://wfoplantlist.org/plant-list),) Leipzig catalogue of vascular plants (LCVP, [https://doi.org/10.1038/s41597-020-00702-z),](https://doi.org/10.1038/s41597-020-00702-z),) and the World Checklist of Vascular Plants (WCVP, [https://powo.science.kew.org).](https://powo.science.kew.org).) Additionally, Asteraceae has a family-specific checklist: the Global Compositae Database([https://www.compositae.org/gcd/index.php).](https://www.compositae.org/gcd/index.php).) A recent paper compared the major plant taxonomy databases: Costa, D. S., Boehnisch, G., Freiberg, M., Govaerts, R., Grenié, M., Hassler, M., Kattge, J., Muellner‐Riehl, A. N., Andrés, B. M. R., Winter, M., Watson, M., Zizka, A. & Wirth, C. (2023). The big four of plant taxonomy -- a comparison of global checklists of vascular plant names. *New Phytologist*. <https://doi.org/10.1111/nph.18961>

\
Estimates for the current number of accepted Asteraceae species range across the databases:

-   TPL = 32,913;

-   GCD = 32,995 -108 hybrids = 32,887;

-   WCVP = 33,994;

-   WFO = 32,095 -312 hybrids = 31,783

With a range of 31,783 to 33,994 accepted Asteraceae species.

Here we use the WCVP to calculate the species richness in each family, since GIFT is standardized against WCVP.

Test database cleaning with Asteraceae. Filter to family -\> filter to the species taxonomic rank -\> only accepted species -\> be sure to remove hybrids.

```{r}
wcvp_asteraceae <- wcvp %>%
  filter(family == "Asteraceae") %>%
  filter(taxon_rank == "Species") %>%
  filter(taxon_status == "Accepted") %>%
  filter(!grepl("×", taxon_name))

distinct(wcvp_asteraceae, taxon_name)
```

For each family with island species, calculate the total number of accepted species (no hybrids) according to WCVP. Note the Adoxaceae - Viburnaceae name mismatch needs to be corrected for.

```{r}
diversity_islands_all <- diversity_islands_all %>%
  mutate(total_species = sapply(family, function(f) {
    if (f == "Adoxaceae") {
      f <- "Viburnaceae"  # Change family name to match wcvp
    }
    wcvp_family <- wcvp %>%
      filter(family == f) %>%
      filter(taxon_rank == "Species") %>%
      filter(taxon_status == "Accepted") %>%
      filter(!grepl("×", taxon_name))
    n_distinct(wcvp_family$taxon_name)
  }))
```

Calculate proportion island species (island species/total species in family). And rename for publication table.

```{r}
diversity_islands_all <- diversity_islands_all %>%
  mutate(prop_island = ((native / total_species)*100) %>% round(digits = 0)) 
```

Repeat with oceanic islands:

```{r}
diversity_islands_oceanic <- diversity_islands_oceanic %>%
  mutate(total_species = sapply(family, function(f) {
    if (f == "Adoxaceae") {
      f <- "Viburnaceae"  # Change family name to match wcvp
    }
    wcvp_family <- wcvp %>%
      filter(family == f) %>%
      filter(taxon_rank == "Species") %>%
      filter(taxon_status == "Accepted") %>%
      filter(!grepl("×", taxon_name))
    n_distinct(wcvp_family$taxon_name)
  }))
```

Calculate proportion island species (island species/total species in family). And rename for publication table.

```{r}
diversity_islands_oceanic <- diversity_islands_oceanic %>%
  mutate(prop_island = ((native / total_species)*100) %>% round(digits = 0))
```

Note there is some mismath between GIFT and WCVP with families: Degeneriaceae, Cynomoriaceae, Mitrastemonaceae, Flagellariaceae, Joinvilleaceae, Myodocarpaceae, Sarcolaenaceae. This does not affect the analysis, but to note with no action needed.

## Save the island diversity data

1\. Full dataset of family diversity. Rename columns with publication ready headers.

```{r}
diversity_islands_all %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../data/data_clean/angio_diversity_islands_all.xlsx")

diversity_islands_oceanic %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../data/data_clean/angio_diversity_islands_oceanic.xlsx")
```

2\. Top ten most diverse families

```{r}
diversity_islands_all_top10 <- diversity_islands_all %>%
  arrange(desc(native)) %>%
    slice(1:10) 

diversity_islands_oceanic_top10 <- diversity_islands_oceanic %>%
  arrange(desc(native)) %>%
    slice(1:10)
```

```{r}
# All island types - top10 (save to cleaned data)
diversity_islands_all_top10  %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../data/data_clean/angio_diversity_islands_all_top10.xlsx")
# All island types - top10 (save to outputs)
diversity_islands_all_top10  %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../outputs/1_Angiosperm_island_diversity/angio_diversity_islands_all_top10.xlsx")


# Oceanic islands - top10 (save to cleaned data)
diversity_islands_oceanic_top10 %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../data/data_clean/angio_diversity_islands_oceanic_top10.xlsx")
# Oceanic islands - top10 (save to outputs)
diversity_islands_oceanic_top10 %>%
  rename(c("Family" = family, "Native island species" = native, "Endemic island species" = endemic, "Total species in family" = total_species, "Proportion insular species" = prop_island)) %>%
  write_xlsx("../outputs/1_Angiosperm_island_diversity/angio_diversity_islands_oceanic_top10.xlsx")
```

# 2. Binomial Tests

(Based on Rampal's suggestions on April 2023 manuscript draft)

Binomial test basics:

-   binom.test(x, n, p)

-   x: number of successes (island asteraceae species)

-   n: number of trials (total number of island angiosperm species)

-   p: probability of success on a given trial (proportion of asteraceae globally)

Get (1) calculation of total angiosperm species on islands and (2) calculation of angiosperm species globally

Calculate the total number of angiosperm species native to islands (all and oceanic)

```{r}
# all islands 
sum(diversity_islands_all$native)
# oceanic islands
sum(diversity_islands_oceanic$native)
```

Calculate the total number of accepted Angiosperm species. WCVP does not make it easy to filter to just angiosperm, so start by getting a list of the families and then manually review.

```{r}
unique_families <- wcvp %>%
  filter(taxon_rank == "Species") %>%
  filter(taxon_status == "Accepted") %>%
  filter(!grepl("×", taxon_name)) %>%
  distinct(family) %>%
  pull(family)
unique_families_df <- as.data.frame(unique_families)
```

Then, manually review this list and identify all the non-angiosperm families:

```{r}
excluded_families <- c(
  "Araucariaceae", "Aspleniaceae", "Cupressaceae", "Cyatheaceae", "Cycadaceae",
  "Cystodiaceae", "Dennstaedtiaceae", "Dipteridaceae", "Ephedraceae", "Equisetaceae",
  "Ginkgoaceae", "Gleicheniaceae", "Gnetaceae", "Hymenophyllaceae", "Isoetaceae",
  "Lindsaeaceae", "Lonchitidaceae", "Lycopodiaceae", "Marattiaceae", "Marsileaceae",
  "Matoniaceae", "Ophioglossaceae", "Osmundaceae", "Pinaceae", "Podocarpaceae",
  "Polypodiaceae", "Psilotaceae", "Pteridaceae", "Saccolomataceae", "Salviniaceae",
  "Schizaeaceae", "Sciadopityaceae", "Selaginellaceae", "Taxaceae", "Welwitschiaceae",
  "Zamiaceae"
)
```

Get a count of the unique angiosperm species:

```{r}
unique_angiosperm_species <- wcvp %>%
  filter(taxon_rank == "Species") %>%
  filter(taxon_status == "Accepted") %>%
  filter(!grepl("×", taxon_name)) %>%
  filter(!family %in% excluded_families) %>%
  distinct(taxon_name) %>%
  summarise(count = n())
unique_angiosperm_species
```

Currently accepted angiosperm species (according to WCVP) = 333,799

Binomial test for Asteraceae:

Global proportion of Asteraceae to all angiosperm species: 33,994 / 333,799 (10%)

Proportion of island Asteraceae to all island angiosperm species: 6,135 / 99,659 (6%)

```{r}
# Asteraceae
binom.test(6135, 99659, .10)
# binom.test(6135, 99659, 10/100, alternative="less")
```

The null hypothesis in the test is that the true probability of success (the proportion of Asteraceae species on islands) is equal to 0.10, which is the overall proportion of Asteraceae to angiosperms globally. The alternative hypothesis is that the true probability of success is not equal to 0.10, indicating a deviation from the expected proportion.

The test result suggests strong evidence against the null hypothesis. The p-value is reported as "\< 2.2e-16," which means the p-value is extremely small and essentially zero. This indicates that the observed number of Asteraceae species on islands significantly deviates from the expected number based on the proportion of 0.10.

The 95 percent confidence interval provides a range estimate for the true probability of success. In this case, the interval is reported as (0.06007543 0.0606999), which means we can be 95 percent confident that the true proportion of Asteraceae species on islands falls within this interval. The sample estimate for the probability of success is reported as 0.06155992, which represents the estimated proportion of Asteraceae species based on the observed data.

Overall, the results suggest that the observed number of Asteraceae species on islands is significantly different from what would be expected based on the overall proportion of Asteraceae to angiosperms globally (0.10). The proportion of Asteraceae species on islands appears to be lower, with a 95 percent confidence interval indicating a range of plausible values.

## Binomial test for the ten most diverse families on (1) all islands (2) oceanic islands

**All islands**

Add the total island species = 99,659 (sum(diversity_islands_all\$native)) and the total number of angiosperms = 333,799 (WCVP).

```{r}
binomial_test_islands_all_10 <- diversity_islands_all_top10 %>%
  select(-c(prop_island, endemic)) %>%
  mutate(total_island_species = 99659, .after = native) %>%
  mutate(total_angiosperm_species = 333799, .after = total_species)
```

Now calculate the percent of each family to the total island species (for the results table) and the overall percent of the family to angiosperms.

```{r}
binomial_test_islands_all_10 <- binomial_test_islands_all_10 %>%
  mutate(global_proportion = total_species / total_angiosperm_species, .after = total_angiosperm_species)
```

Run a binomial for each family.

```{r}
binomial_test_islands_all_10 <- binomial_test_islands_all_10 %>%
  group_by(family) %>%
  mutate(binom = list(broom::tidy(binom.test(native, total_island_species, global_proportion)))) %>%
  # expand the binom.test data frame
  unnest(cols = c(binom)) %>%
  # tidy up:
  # convert proportions to percent 
  mutate(global_proportion = (global_proportion * 100)%>% round(digits = 1)) %>% 
  mutate(estimate = (estimate * 100)%>% round(digits = 2)) %>% 
  mutate(conf.low = (conf.low * 100)%>% round(digits = 2)) %>% 
  mutate(conf.high = (conf.high * 100)%>% round(digits = 2)) %>% 
  ## Only show p-values over 0.001 (those under report as <0.001)
  mutate(p.value = ifelse(p.value < 0.001, 
                           "<0.001", 
                           as.character(round(p.value, 3)))) %>% 
  # select the columns to keep
  select(
    family, 
    native,
    total_island_species, 
    global_proportion, 
    p.value,
    estimate, 
    conf.low, 
    conf.high
  )
```

**Oceanic islands**

Add the total oceanic island species = 23853 (sum(diversity_islands_oceanic\$native)) and the total number of angiosperms = 333,799 (WCVP).

```{r}
binomial_test_islands_oceanic_10 <- diversity_islands_oceanic_top10 %>%
  select(-c(prop_island, endemic)) %>%
  mutate(total_island_species = 23853, .after = native) %>%
  mutate(total_angiosperm_species = 333799, .after = total_species)
```

Now calculate the overall percent of the family to angiosperms.

```{r}
binomial_test_islands_oceanic_10 <- binomial_test_islands_oceanic_10 %>%
  mutate(global_proportion = total_species / total_angiosperm_species, .after = total_angiosperm_species)
```

Run a binomial for each family.

```{r}
binomial_test_islands_oceanic_10 <- binomial_test_islands_oceanic_10 %>%
  group_by(family) %>%
  mutate(binom = list(broom::tidy(binom.test(native, total_island_species, global_proportion)))) %>%
  # expand the binom.test data frame
  unnest(cols = c(binom)) %>%
  # tidy up:
  # convert proportions to percent 
  mutate(global_proportion = (global_proportion * 100)%>% round(digits = 1)) %>% 
  mutate(estimate = (estimate * 100)%>% round(digits = 2)) %>% 
  mutate(conf.low = (conf.low * 100)%>% round(digits = 2)) %>% 
  mutate(conf.high = (conf.high * 100)%>% round(digits = 2)) %>% 
  ## Only show p-values over 0.001 (those under report as <0.001)
  mutate(p.value = ifelse(p.value < 0.001, 
                           "<0.001", 
                           as.character(round(p.value, 3)))) %>% 
  # select the columns to keep
  select(
    family, 
    native,
    total_island_species, 
    global_proportion, 
    p.value,
    estimate, 
    conf.low, 
    conf.high
  )
```

Save the datatables to the outputs directory

```{r}
# all islands
binomial_test_islands_all_10 %>%
  rename(c("Family" = family, 
           "Native island species (x)" = native, 
           "Angiosperm species on islands (n)" = total_island_species, 
           "Proportion of family globally (p)" = global_proportion, 
           "p-value" = p.value,
           "Sample estimate" = estimate,
           "Confidence interval (low)" = conf.low,
           "Confidence interval (high)" = conf.high)) %>%
  write_xlsx("../outputs/1_Angiosperm_island_diversity/binomial_test_islands_all_10.xlsx")

# oceanic islands
binomial_test_islands_oceanic_10 %>%
  rename(c("Family" = family, 
           "Native island species (x)" = native, 
           "Angiosperm species on islands (n)" = total_island_species, 
           "Proportion of family globally (p)" = global_proportion, 
           "p-value" = p.value,
           "Sample estimate" = estimate,
           "Confidence interval (low)" = conf.low,
           "Confidence interval (high)" = conf.high)) %>%
  write_xlsx("../outputs/1_Angiosperm_island_diversity/binomial_test_islands_oceanic_10.xlsx")
```

# 3. Null models and island simulations

Set up null model and define effect size.

Set the null to 10,000

```{r}
n_null<-10000
```

Empirical probability Matthews et al 2020. Define effect size: Equation taken from Matthews et al 2022 Island bird review JBI

```{r}
effect_size<-function(null,obs) 
    {
    null<-c(null,obs)
    p_val<-(length(which(null<obs))+
              (length(which(null==obs))/2))/(length(null)+1)
    ES<-probitlink(p_val)
    
    print("Probability  Effect size")
    return(c(p_val,ES))
}
```

## Simulate island community (All islands)

```{r}
# convert to dataframe
fams <- as.data.frame(diversity_islands_all_top10)

total_spec<-sum(fams[,4])
total_island<-sum(fams[,2])
```

Create global pool of the top ten families overall species diversity.

comm size = total number of species for the top 10 families on islands.

```{r}
comm<-c(rep(fams[1,1],fams[1,4]),
        rep(fams[2,1],fams[2,4]),
        rep(fams[3,1],fams[3,4]),
        rep(fams[4,1],fams[4,4]),
        rep(fams[5,1],fams[5,4]),
        rep(fams[6,1],fams[6,4]),
        rep(fams[7,1],fams[7,4]),
        rep(fams[8,1],fams[8,4]),
        rep(fams[9,1],fams[9,4]),
        rep(fams[10,1],fams[10,4]))
```

Randomly sample from global pool to create island pools with the same number of species as the total island pool (families selected at random).

```{r}
tr_table<-c()

for(i in 1:n_null){
  
tr<-sample(comm,total_island,replace = FALSE)
tr_table_new<-t(table(tr))
tr_table<-rbind(tr_table,tr_table_new)
}
```

Create tibble for plots later.

```{r}
iteration<-1:n_null
tr_table_tibble<-add_column(as_tibble(tr_table),iteration,.before = T)
tr_table_tibble <-
  pivot_longer(tr_table_tibble,
               cols = 2:11,
               names_to ="family",
               values_to = "species")
```

Histograms just to visualize results

```{r}
par(mfrow=c(10,1),
    mai = c(0.3, 0.5, 0.1, 0.2),
    oma=c(0,0.5,0.1,0),
    mgp=c(2, 1, 0))
```

```{r}
hist(as.vector(tr_table[,"Orchidaceae"]),breaks=100,xlim=c(0,12000),
     main='Orchidaceae', xlab=NA)
abline(v=fams[1,2],col='red')
abline(v=quantile(tr_table[,"Orchidaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Rubiaceae"]),breaks=100,xlim=c(0,10000),
     main='Rubiaceae', xlab=NA)
abline(v=fams[2,2],col='red')
abline(v=quantile(tr_table[,"Rubiaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Asteraceae"]),breaks=100,xlim=c(0,10000),
     main='Asteraceae', xlab=NA)
abline(v=fams[3,2],col='red')
abline(v=quantile(tr_table[,"Asteraceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Fabaceae"]),breaks=100,xlim=c(0,10000),
     main='Fabaceae', xlab="Number of island species")
abline(v=fams[4,2],col='red')
abline(v=quantile(tr_table[,"Fabaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Poaceae"]),breaks=100,xlim=c(0,10000),
     main='Poaceae', xlab="Number of island species")
abline(v=fams[5,2],col='red')
abline(v=quantile(tr_table[,"Poaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Myrtaceae"]),breaks=100,xlim=c(0,10000),
     main='Myrtaceae', xlab="Number of island species")
abline(v=fams[6,2],col='red')
abline(v=quantile(tr_table[,"Myrtaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Cyperaceae"]),breaks=100,xlim=c(0,10000),
     main='Cyperaceae', xlab="Number of island species")
abline(v=fams[7,2],col='red')
abline(v=quantile(tr_table[,"Cyperaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Apocynaceae"]),breaks=100,xlim=c(0,10000),
     main='Apocynaceae', xlab="Number of island species")
abline(v=fams[8,2],col='red')
abline(v=quantile(tr_table[,"Apocynaceae"],c(0.025,0.975)),lty=2)


hist(as.vector(tr_table[,"Euphorbiaceae"]),breaks=100,xlim=c(0,10000),
     main='Euphorbiaceae', xlab="Number of island species")
abline(v=fams[9,2],col='red')
abline(v=quantile(tr_table[,"Euphorbiaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table[,"Melastomataceae"]),breaks=100,xlim=c(0,10000),
     main='Melastomataceae', xlab="Number of island species")
abline(v=fams[10,2],col='red')
abline(v=quantile(tr_table[,"Melastomataceae"],c(0.025,0.975)),lty=2)
```

Check for significance/effect size using Matthews et al. metric

```{r}
null<-as.vector(tr_table[,"Orchidaceae"])
obs<-fams[1,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Rubiaceae"])
obs<-fams[2,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Asteraceae"])
obs<-fams[3,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Fabaceae"])
obs<-fams[4,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Poaceae"])
obs<-fams[5,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Myrtaceae"])
obs<-fams[6,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Cyperaceae"])
obs<-fams[7,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Apocynaceae"])
obs<-fams[8,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Euphorbiaceae"])
obs<-fams[9,2]
effect_size(null,obs)

null<-as.vector(tr_table[,"Melastomataceae"])
obs<-fams[10,2]
effect_size(null,obs)


```

## Simulate island community (Oceanic islands)

```{r}
# convert to dataframe
oceanic_fams <- as.data.frame(diversity_islands_oceanic_top10)

total_spec<-sum(oceanic_fams[,4])
total_island<-sum(oceanic_fams[,2])
```

Create global pool of the top ten families overall species diversity comm size = total number of species for the top 10 families (on oceanic islands).

```{r}
comm<-c(rep(oceanic_fams[1,1],oceanic_fams[1,4]),
        rep(oceanic_fams[2,1],oceanic_fams[2,4]),
        rep(oceanic_fams[3,1],oceanic_fams[3,4]),
        rep(oceanic_fams[4,1],oceanic_fams[4,4]),
        rep(oceanic_fams[5,1],oceanic_fams[5,4]),
        rep(oceanic_fams[6,1],oceanic_fams[6,4]),
        rep(oceanic_fams[7,1],oceanic_fams[7,4]),
        rep(oceanic_fams[8,1],oceanic_fams[8,4]),
        rep(oceanic_fams[9,1],oceanic_fams[9,4]),
        rep(oceanic_fams[10,1],oceanic_fams[10,4]))
```

Randomly sample from global pool to create oceanic island pools with the same number of species as the total oceanic island pool.

```{r}
tr_table_oceanic<-c()

for(i in 1:n_null){
  
  tr<-sample(comm,total_island,replace = FALSE)
  tr_table_new<-table(tr)
  tr_table_oceanic<-rbind(tr_table_oceanic,tr_table_new)
}
```

Create tibble for plots later.

```{r}
iteration<-1:n_null
tr_table_oceanic_tibble<-add_column(as_tibble(tr_table_oceanic),iteration,.before = T)
tr_table_oceanic_tibble<-
  pivot_longer(tr_table_oceanic_tibble,
               cols = 2:11,
               names_to ="family",
               values_to = "species")
```

Histograms just to visualize results.

```{r}
par(mfrow=c(4,1),
    mai = c(0.3, 0.5, 0.1, 0.2),
    oma=c(0,0.5,0.1,0),
    mgp=c(2, 1, 0))
```

```{r}
hist(as.vector(tr_table_oceanic[,"Asteraceae"]),breaks=100,xlim=c(0,4000),
     main='Asteraceae', xlab=NA)
abline(v=oceanic_fams[1,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Asteraceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Orchidaceae"]),breaks=100,xlim=c(0,4000),
     main='Orchidaceae', xlab=NA)
abline(v=oceanic_fams[2,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Orchidaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Poaceae"]),breaks=100,xlim=c(0,4000),
     main='Poaceae', xlab="Number of island species")
abline(v=oceanic_fams[3,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Poaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Rubiaceae"]),breaks=100,xlim=c(0,4000),
     main='Rubiaceae', xlab=NA)
abline(v=oceanic_fams[4,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Rubiaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Fabaceae"]),breaks=100,xlim=c(0,4000),
     main='Fabaceae', xlab=NA)
abline(v=oceanic_fams[5,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Fabaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Cyperaceae"]),breaks=100,xlim=c(0,4000),
     main='Cyperaceae', xlab=NA)
abline(v=oceanic_fams[6,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Cyperaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Euphorbiaceae"]),breaks=100,xlim=c(0,4000),
     main='Euphorbiaceae', xlab=NA)
abline(v=oceanic_fams[7,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Euphorbiaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Myrtaceae"]),breaks=100,xlim=c(0,4000),
     main='Myrtaceae', xlab=NA)
abline(v=oceanic_fams[8,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Myrtaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Malvaceae"]),breaks=100,xlim=c(0,4000),
     main='Malvaceae', xlab=NA)
abline(v=oceanic_fams[9,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Malvaceae"],c(0.025,0.975)),lty=2)

hist(as.vector(tr_table_oceanic[,"Apocynaceae"]),breaks=100,xlim=c(0,4000),
     main='Apocynaceae', xlab=NA)
abline(v=oceanic_fams[10,2],col='red')
abline(v=quantile(tr_table_oceanic[,"Apocynaceae"],c(0.025,0.975)),lty=2)


```

Check for significance/effect size using Matthews et al. metric

```{r}
null<-as.vector(tr_table_oceanic[,"Asteraceae"])
obs<-oceanic_fams[1,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Orchidaceae"])
obs<-oceanic_fams[2,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Poaceae"])
obs<-oceanic_fams[3,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Rubiaceae"])
obs<-oceanic_fams[4,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Fabaceae"])
obs<-oceanic_fams[5,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Cyperaceae"])
obs<-oceanic_fams[6,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Euphorbiaceae"])
obs<-oceanic_fams[7,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Myrtaceae"])
obs<-oceanic_fams[8,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Malvaceae"])
obs<-oceanic_fams[9,2]
effect_size(null,obs)

null<-as.vector(tr_table_oceanic[,"Apocynaceae"])
obs<-oceanic_fams[10,2]
effect_size(null,obs)


```

# 4. Results: Visualize insular diversity and null expectation for angiosperms

## Top ten families with the most native and endemic species on islands

Plot the top 10 most diverse angiosperm families on across all islands.

```{r}
angio_diversity_islands_all <- diversity_islands_all %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native), color="black") +
  geom_point( aes(x=family, y=endemic), color="#009999", size=3 ) +
  geom_point( aes(x=family, y=native), color="#003399", size=3 ) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity") +
  ggtitle("Ten most diverse angiosperm families on islands")


angio_diversity_islands_all
```

Plot the top 10 most diverse angiosperm families on across oceanic islands.

```{r}
angio_diversity_islands_oceanic <- diversity_islands_oceanic %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native), color="black") +
  geom_point( aes(x=family, y=endemic), color="#FF9933", size=3 ) +
  geom_point( aes(x=family, y=native), color="#b12a90", size=3 ) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity") +
  ggtitle("Ten most diverse angiosperm families on oceanic islands")


angio_diversity_islands_oceanic
```

Combine into one plot.

```{r}
# All islands
angio_div_islands_all <- diversity_islands_all %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native)) +
  geom_point( aes(x=family, y=endemic), color="#009999", size=3 ) +
  geom_point( aes(x=family, y=native), color="#003399", size=3) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity") +
  ggtitle("All islands")+ 
  theme(plot.title = element_text(hjust = 0.5))
angio_div_islands_all


# oceanic
angio_div_islands_oceanic <- diversity_islands_oceanic %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native)) +
  geom_point( aes(x=family, y=endemic), color="#FF9933", size=3 ) +
  geom_point( aes(x=family, y=native), color="#b12a90", size=3) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity") +
  ggtitle("Oceanic islands")+ 
  theme(plot.title = element_text(hjust = 0.5))
angio_div_islands_oceanic

# Combine plots with cowplot 

# first combine plots
angio_div <- plot_grid(angio_div_islands_all, angio_div_islands_oceanic, labels = "AUTO")
angio_div

# then add a title
title <- ggdraw() + 
  draw_label(
    "Most diverse angiosperm families on islands",
    fontface = 'bold',
    x = 0,
    hjust = 0) +
  theme(
    # add margin on the left of the drawing canvas,
    # so title is aligned with left edge of first plot
    plot.margin = margin(0, 0, 0, 0)
  )

# Add title and plots together
angio_diversity_islands <- plot_grid(
  title, angio_div,
  ncol = 1,
  # rel_heights values control vertical title margins
  rel_heights = c(0.1, 1)
)
angio_diversity_islands
```

```{r}
ggsave("../outputs/1_Angiosperm_island_diversity/Angio_diversity_islands.svg", angio_diversity_islands, width = 200, height = 100, units = "mm")
```

## Observed island diversity compared to the null expectation

Plot the observed number of native species compared to null expectation (All islands)

```{r}
# set the order
fams<-fams[order(fams$native),]
```

```{r}
angio_null_islands_all <- ggplot(tr_table_tibble, aes(x=factor(family,level=fams$family),
                                  y=species)) + 
  stat_histinterval(point_size=0.1, scale = 0.95, interval_side = 0.5, interval_color = "darkgrey", point_color = "darkgrey") +
  geom_point(data=fams, mapping = aes(x = family, y = native), 
             shape = 19, size = 2,
             # color is hard coded here...so check/update if needed (Note! name in reverse order)
             colour = c("black", 
                        "black", 
                        "black", 
                        "black", 
                        "black",
                        "black", 
                        "#cc0000", 
                        "#cc0000", 
                        "black", 
                        "black" )) +
  coord_flip() +
  theme_bw()+
  ggtitle("All islands") +
  xlab(NULL) +
  ylab(NULL) 

# update the theme to bold Asteraceae and center title
Aster_bold <- ifelse(fams$family == "Asteraceae","bold","plain")
angio_null_islands_all <- angio_null_islands_all +
  theme(
    plot.title = element_text(hjust = 0.5),
    axis.text.y=element_text(face=Aster_bold))
angio_null_islands_all
```

Plot the observed number of native species compared to null expectation (Oceanic islands)

```{r}
oceanic_fams<-oceanic_fams[order(oceanic_fams$native),]
```

```{r}
angio_null_islands_oceanic <- ggplot(tr_table_oceanic_tibble, aes(x=factor(family,level=oceanic_fams$family), y=species)) + 
  stat_histinterval(point_size=0.1, scale = 0.95, interval_side = 0.5, interval_color = "darkgrey", point_color = "darkgrey") +
  geom_point(data=oceanic_fams, mapping = aes(x = family, y = native), 
             shape = 19, size = 2,
             colour = c("black", 
                        "black", 
                        "black", 
                        "black", 
                        "black",
                        "#cc0000", 
                        "black", 
                        "black", 
                        "#cc0000", 
                        "#cc0000")) + 
  coord_flip() +
  theme_bw()+
  ggtitle("Oceanic islands") +
  xlab(NULL) +
  ylab(NULL) 
angio_null_islands_oceanic 
# update the theme to bold Asteraceae and center title
Aster_bold_o <- ifelse(oceanic_fams$family == "Asteraceae","bold","plain")
angio_null_islands_oceanic  <- angio_null_islands_oceanic  +
  theme(
    plot.title = element_text(hjust = 0.5),
    axis.text.y=element_text(face=Aster_bold_o))
angio_null_islands_oceanic 
```

Combine the angiosperm island diversity plots with the null expectations plots into one 4-panel plot.

```{r}
# Redraw the plots with new variables to tune specifically for the combined 4 plots 

# Angiosperm diversity: All islands
p1 <- diversity_islands_all %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native)) +
  geom_point( aes(x=family, y=endemic), color="#009999", size=3 ) +
  geom_point( aes(x=family, y=native), color="#003399", size=3) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity")
p1

# Null expectations: All islands
p2 <- ggplot(tr_table_tibble, aes(x=factor(family,level=fams$family),
                                  y=species)) + 
  stat_histinterval(point_size=0.1, scale = 0.95, interval_side = 0.5, interval_color = "darkgrey", point_color = "darkgrey") +
  geom_point(data=fams, mapping = aes(x = family, y = native), 
             shape = 19, size = 2,
             # color is hard coded here...so check/update if needed (Note! name in reverse order)
             colour = c("black", 
                        "black", 
                        "black", 
                        "black", 
                        "black",
                        "black", 
                        "#cc0000", 
                        "#cc0000", 
                        "black", 
                        "black" )) +
  coord_flip() +
  theme_bw()+
  theme(axis.text.y=element_blank(), 
        axis.ticks.y=element_blank()) +
  xlab(NULL) +
  ylab("Species diversity") 
p2

# Angiosperm diversity: Oceanic islands
p3 <- diversity_islands_oceanic %>%
  mutate(family = fct_reorder(family, native)) %>%
  slice_head(n = 10) %>%
  ggplot() +
  geom_segment( aes(x=family, xend=family, y=endemic, yend=native)) +
  geom_point( aes(x=family, y=endemic), color="#FF9933", size=3 ) +
  geom_point( aes(x=family, y=native), color="#b12a90", size=3) +
  coord_flip()+
  theme_bw() +
  #theme(
  #  legend.position = "none")+
  xlab("") +
  ylab("Species diversity")
p3

p4 <- ggplot(tr_table_oceanic_tibble, aes(x=factor(family,level=oceanic_fams$family), y=species)) + 
  stat_histinterval(point_size=0.1, scale = 0.95, interval_side = 0.5, interval_color = "darkgrey", point_color = "darkgrey") +
  geom_point(data=oceanic_fams, mapping = aes(x = family, y = native), 
             shape = 19, size = 2,
             colour = c("black", 
                        "black", 
                        "black", 
                        "black", 
                        "black",
                        "#cc0000", 
                        "black", 
                        "black", 
                        "#cc0000", 
                        "#cc0000")) + 
  coord_flip() +
  theme_bw()+
  theme(axis.text.y=element_blank(), 
        axis.ticks.y=element_blank()) +
  xlab(NULL) +
  ylab("Species diversity") 
p4



angio_div_plots <- (p1 + p2) / (p3 + p4)
angio_div_plots
```

Save the plot. For publication, do minor edits: bold Asteraceae, create a legend for each panel, change the font (Gill).

```{r}
ggsave("../outputs/1_Angiosperm_island_diversity/Angio_diversity_islands_simulations_raw.svg", angio_div_plots, width = 150, height = 125, units = "mm")
```
