==> code/2_B_Intrafamily_island_diversity.Rmd <==
---
title: "Intrafamily island diversity"
output: html_document
date: "2023-09-25"
---

# Overview

Aim: Identify the most diverse clades within Asteraceae for island species richness. Is island diversity unevenly distributed within the family? Is island diversity over-represented in any lineages of the family?

Main steps (Outline):

1.  Calculate diversity for each major lineage (at the tribe level) in Asteraceae
    1.  Calculate the overall number of species (globally) in each tribe

    2.  Calculate the number of island species (native and endemic) in each tribe

    3.  Build table and save
2.  Binomial tests for the tribes
3.  Simulations of null expectations for tribes

## Preparation

Load libraries:

```{r}
library(tidyverse)
library(readxl)
library(writexl)
library(ggdist)
library(cowplot)
library(svglite)
```

Load data:

Read in the Asteraceae taxonomy (`Asteraceae_taxonomy_accepted.xlsx`). This is the family-wide taxonomy for accepted species using the Global Compositae Database backbone and built in the `0_Data_preparation.Rmd` script.

```{r}
Asteraceae_taxonomy_accepted <- read_excel("../data/data_clean/asteraceae_taxonomy_accepted.xlsx", guess_max = 30000)
```

Read in the Asteraceae island checklists (`Asteraceae_islands_checklists.xlsx`) which have the GCD taxonomy (status, tribe, subfamily) integrated.

```{r}
Asteraceae_islands_checklists <- read_excel("../data/data_clean/Asteraceae_islands_checklists.xlsx", guess_max = 10000)
```

Read in the phylogenetic order for Subfamilies based on Mandel 2019. This is simplify for organization of the order in the output datatables.

```{r}
phylo_order <- read_xlsx("../data/data_raw/Mandel_phylo_order.xlsx")
```

# Calculate diversity for each major lineage (at the tribe level) in Asteraceae

## Calculate the overall number of species (globally) in each tribe

Note there are 41 species that do not have a tribe+subfamily assessment. e

```{r}
na_tribes <- Asteraceae_taxonomy_accepted %>%
  filter(is.na(Tribe))
```

These are mostly unplaced (Callilepis) genera and typos (Cosmus). These are not island species will have a very minor impact overall number of species per tribe, so we will note and filter out.

```{r}
Asteraceae_taxonomy_accepted <- Asteraceae_taxonomy_accepted %>%
  filter(!is.na(Tribe))
```

Calculate overall diversity per tribe:

```{r}
Tribe_diversity_global <- Asteraceae_taxonomy_accepted %>%
  group_by(Tribe, Subfamily) %>%
  count(Tribe, name = "Diversity_global")
```

## Calculate the number of island species (native and endemic) in each tribe

Calculate the number of island native species in each tribe.

```{r}
Tribe_diversity_island_natives <- Asteraceae_islands_checklists %>%
  distinct(work_species, .keep_all = TRUE) %>%
  group_by(Tribe, Subfamily) %>%
  count(Tribe, name = "Diversity_island_natives")
```

Calculate the number of island endemic species in each tribe.

```{r}
Tribe_diversity_island_endemics <- Asteraceae_islands_checklists %>%
  filter(endemic_list == 1) %>%
  distinct(work_species, .keep_all = TRUE) %>%
  group_by(Tribe, Subfamily) %>%
  count(Tribe, name = "Diversity_island_endemics")
```

## Build table and save

Now create the master table with the total tribal diversity.

```{r}
# Join the global diversity table to the island natives table
Intrafamily_diversity <- left_join(
  Tribe_diversity_global,
  Tribe_diversity_island_natives,
  by = c("Tribe", "Subfamily"))

# Join the island endemics to intrafamily diversity table 
Intrafamily_diversity <- left_join(
  Intrafamily_diversity,
  Tribe_diversity_island_endemics,
  by = c("Tribe", "Subfamily")) %>%
  replace(is.na(.), 0) # replace nas with 0
```

Calculate the proportion of island species within each tribe.

```{r}
Intrafamily_diversity <- Intrafamily_diversity %>%
  mutate(Proportion_island = ((Diversity_island_natives/Diversity_global)*100) %>% round(digits = 0))
```

Add in the phylo order for organization.

```{r}
Intrafamily_diversity <- left_join(
  Intrafamily_diversity,
  phylo_order,
  by = "Subfamily")

Intrafamily_diversity <- Intrafamily_diversity %>% 
  relocate(phylo_order, .before = Tribe)
```

Save the Intrafamily datatable in the data_clean directory and the outputs directory with publication ready column headers

```{r}
write_xlsx(Intrafamily_diversity, "../data/data_clean/Intrafamily_diversity.xlsx")

# Save to outputs with nicely formatted column headers
Intrafamily_diversity %>%
  rename(c("Total species in tribe" = Diversity_global,
           "Native island species" = Diversity_island_natives,
           "Endemic island species" = Diversity_island_endemics, 
           "Proportion island species" = Proportion_island)) %>%
  write_xlsx("../outputs/2_Diversity_patterns/Intrafamily_diversity.xlsx")
```

## Results: Visualize intrafamily diversity

First, we create a preliminary bar chart of the diversity of island species within each tribe.

```{r}
# define th colors for the subfamilies
subfamily_colors <- c(Stifftioideae = "#92278F", Mutisioideae = "#BE202F", Gochnatioideae ="#F15A2B", Pertyoideae = "#F79320", Tarchonanthoideae = "#FBB042", Dicomoideae = "#FDD25C", Carduoideae = "#8EC740", Vernonioideae = "#49A347", Cichorioideae = "#02818B", Asteroideae = "#01604C")
```

Create factors for the plotting.

```{r}
Intrafamily_diversity_fct <- Intrafamily_diversity %>%
mutate(Tribe = factor(Tribe, levels = c("Barnadesieae", "Famatinantheae", "Stifftieae", "Hyalideae", "Mutisieae", "Nassauvieae", "Onoserideae", "Gochnatieae", "Wunderlichieae", "Cyclolepis", "Hecastocleideae", "Pertyeae", "Tarchonantheae", "Oldenburgieae", "Dicomeae", "Cardueae", "Gymnarrheneae", "Vernonieae", "Liabeae", "Arctotideae", "Distephaneae", "Platycarpheae", "Moquinieae", "Eremothamneae", "Cichorieae", "Corymbieae", "Astereae", "Senecioneae", "Gnaphalieae", "Anthemideae", "Eupatorieae", "Heliantheae", "Inuleae", "Coreopsideae", "Madieae", "Tageteae", "Millerieae", "Athroismeae", "Neurolaeneae", "Calenduleae", "Perityleae", "Helenieae", "Bahieae", "Doroniceae", "Chaenactideae", "Polymnieae", "Feddeeae")))  %>%
mutate(Subfamily = factor(Subfamily, levels =c("Barnadesioideae", "Famatinanthoideae", "Stifftioideae", "Mutisioideae", "Gochnatioideae", "Wunderlichioideae", "Hecastocleidoideae", "Pertyoideae", "Tarchonanthoideae", "Dicomoideae", "Carduoideae", "Gymnarrhenoideae", "Vernonioideae", "Cichorioideae", "Corymbioideae", "Asteroideae")))

```

```{r}
Tribe_island_diversity <- Intrafamily_diversity_fct %>%
  filter(Diversity_island_natives != 0) %>%
  ggplot(
       aes(x = Tribe, fill = Subfamily)) +
  geom_bar(aes(y = Diversity_island_natives), stat = "identity", alpha=0.60) +
  geom_bar(aes(y = Diversity_island_endemics), stat = "identity") +
  scale_fill_manual(values = subfamily_colors) +
  xlab("Tribe") + ylab("Number of island species") + labs(fill = "Subfamily") +
  theme_bw() +
  theme(axis.text.x=element_text(size=10, angle=30, hjust = 1), # adjust the angle and font of x labels
        legend.title = element_text(size =10), # adjust the legend
        legend.text = element_text(size = 8))

Tribe_island_diversity
```

The main output is the diversity (global and island natives) plotted along side the major lineages on the family-level phylogeny. To do this we will use the Mandel et al. 2019 phylogeny (<https://doi.org/10.1073/pnas.1903871116>). Plotting done in Adobe Illustrator in file `Intrafamily_diversity_phylogey.ai`. Below is the final figure:

# Binomial tests for the tribes

Binomial test basics:

-   binom.test(x, n, p)

-   x: number of successes (island species in tribe)

-   n: number of trials (total number of native island species in family)

-   p: probability of success on a given trial (proportion of tribe in overall Asteraceae diversity)

Get (1) calculation total number of native island species and (2) calculation of overall accepted species in Asteraceae

```{r}
n_distinct(Asteraceae_islands_checklists$work_species)
```

We know the number of accepted species in Asteraceae from the WCVP calculated in `1_Angiosperm_island_diversity.Rmd` = 33,994.

Add the total number of island species = 6135 and the total number of Asteraceae species = 33994

```{r}
binomial_test_intrafamily_diversity <- Intrafamily_diversity %>%
  filter(Diversity_island_natives != 0) %>%
  select(-c(Diversity_island_endemics, Proportion_island)) %>%
  mutate(Total_island_species = 6135, .after = Diversity_island_natives) %>%
  mutate(Total_asteraceae_species = 33994, .after = Total_island_species)
```

Next calculate the percent of each family to the total island species (for the results table) and the overall percent of the family to angiosperms.

```{r}
binomial_test_intrafamily_diversity <- binomial_test_intrafamily_diversity %>%
  mutate(Tribe_proportion = Diversity_global / Total_asteraceae_species, .after = Total_island_species)
```

Run Binomial test for each tribe

```{r}
binomial_test_intrafamily_diversity <- binomial_test_intrafamily_diversity %>%
  group_by(Tribe) %>%
  mutate(binom = list(broom::tidy(binom.test(Diversity_island_natives, Total_island_species, Tribe_proportion)))) %>%
  # expand the binom.test data frame
  unnest(cols = c(binom)) %>%
  # tidy up:
  # convert proportions to percent 
  mutate(Tribe_proportion = (Tribe_proportion * 100)%>% round(digits = 1)) %>% 
  mutate(estimate = (estimate * 100)%>% round(digits = 2)) %>% 
  mutate(conf.low = (conf.low * 100)%>% round(digits = 2)) %>% 
  mutate(conf.high = (conf.high * 100)%>% round(digits = 2)) %>% 
  ## Only show p-values over 0.001 (those under report as <0.001)
  mutate(p.value = ifelse(p.value < 0.001, 
                           "<0.001", 
                           as.character(round(p.value, 3)))) %>% 
  # select the columns to keep
  select(
    Tribe, 
    Subfamily,
    Diversity_island_natives,
    Total_island_species, 
    Tribe_proportion, 
    p.value,
    estimate, 
    conf.low, 
    conf.high
  )
```

Add in the phylo order for organization.

```{r}
binomial_test_intrafamily_diversity <- left_join(
  binomial_test_intrafamily_diversity,
  phylo_order,
  by = "Subfamily")

binomial_test_intrafamily_diversity <- binomial_test_intrafamily_diversity %>% 
  relocate(phylo_order, .before = Tribe)
```

Save the binomial test table to `../Outputs/2_Diversity_patterns`.

```{r}
binomial_test_intrafamily_diversity %>%
  rename(c("Native island species (x)" = Diversity_island_natives,
           "Aster. species on islands (n)" = Total_island_species,
           "Proportion of tribe in Aster. (p)" = Tribe_proportion,
           "p-value" = p.value,
           "Sample estimate" = estimate,
           "Confidence interval (low)" = conf.low,
           "Confidence interval (high)" = conf.high)) %>%
  write_xlsx("../outputs/2_Diversity_patterns/Binomial_test_intrafamily_diversity.xlsx")
```

# Simulations of null expectations

Simulate tribes and get a distribution of the null expectation which can be compared to the observed number of island species in each tribe.

Set the null to 10,000

```{r}
n_null<-10000
```

Define the total size of the family (note: here based on GCD) and the total number of native island species.

```{r}
# filter to only tribes with island species
intrafam <- Intrafamily_diversity %>%
  filter(Diversity_island_natives != 0)

# convert to dataframe
intrafam <- as.data.frame(intrafam) 

total_spec<-sum(intrafam[,4])
total_island<-sum(intrafam[,5])
```

Create a pool of Asteraceae global diversity

comm size = the total number of species in Asteraceae (according to GCD)

```{r}
comm<-c(rep(intrafam[1,2],intrafam[1,4]),
        rep(intrafam[2,2],intrafam[2,4]),
        rep(intrafam[3,2],intrafam[3,4]),
        rep(intrafam[4,2],intrafam[4,4]),
        rep(intrafam[5,2],intrafam[5,4]),
        rep(intrafam[6,2],intrafam[6,4]),
        rep(intrafam[7,2],intrafam[7,4]),
        rep(intrafam[8,2],intrafam[8,4]),
        rep(intrafam[9,2],intrafam[9,4]),
        rep(intrafam[10,2],intrafam[10,4]),
        rep(intrafam[11,2],intrafam[11,4]),
        rep(intrafam[12,2],intrafam[12,4]),
        rep(intrafam[13,2],intrafam[13,4]),
        rep(intrafam[14,2],intrafam[14,4]),
        rep(intrafam[15,2],intrafam[15,4]),
        rep(intrafam[16,2],intrafam[16,4]),
        rep(intrafam[17,2],intrafam[17,4]),
        rep(intrafam[18,2],intrafam[18,4]),
        rep(intrafam[19,2],intrafam[19,4]),
        rep(intrafam[20,2],intrafam[20,4]),
        rep(intrafam[21,2],intrafam[21,4]),
        rep(intrafam[22,2],intrafam[22,4]),
        rep(intrafam[23,2],intrafam[23,4]),
        rep(intrafam[24,2],intrafam[24,4]),
        rep(intrafam[25,2],intrafam[25,4]),
        rep(intrafam[26,2],intrafam[26,4]),
        rep(intrafam[27,2],intrafam[27,4]),
        rep(intrafam[28,2],intrafam[28,4]),
        rep(intrafam[29,2],intrafam[29,4]),
        rep(intrafam[30,2],intrafam[30,4]),
        rep(intrafam[31,2],intrafam[31,4]),
        rep(intrafam[32,2],intrafam[32,4]),
        rep(intrafam[33,2],intrafam[33,4]))
```

Randomly sample from the family pool of Asteraceae (`comm`) to create island pools with the same number of species as the total number of native Aster on islands (`total_island`) (tribes selected at random).

```{r}
# Initialize an empty matrix with the correct number of columns
unique_comm <- unique(comm)
num_unique_comm <- length(unique_comm)
tr_table <- matrix(0, nrow = n_null, ncol = num_unique_comm)
colnames(tr_table) <- unique_comm

for (i in 1:n_null) {
  tr <- sample(comm, total_island, replace = FALSE)
  tr_table_new <- table(tr)
  
  # Ensure that tr_table_new has the correct column names
  tr_table_new <- tr_table_new[names(tr_table_new) %in% unique_comm]
  
  # Add the counts to the corresponding columns in the tr_table matrix
  tr_table[i, names(tr_table_new)] <- as.numeric(tr_table_new)
}

# Convert the matrix to a data frame
tr_table <- as.data.frame(tr_table)
```

```{r}
iteration<-1:n_null
tr_table_tibble<-add_column(as_tibble(tr_table),iteration,.before = T)
tr_table_tibble <-
  pivot_longer(tr_table_tibble,
               cols = 2:34,
               names_to ="tribe",
               values_to = "species")
```

Plot to compare observed diversity to null expectation

```{r}
# Reorder the tribe factor based on phylo_order
tr_table_tibble$tribe <- factor(tr_table_tibble$tribe, level = intrafam$Tribe[order(-intrafam$phylo_order)])

tribe_null_expectation <- ggplot(tr_table_tibble, aes(x = tribe, y = species))+ 
  stat_histinterval(point_size=0.1, scale = 0.95, interval_size = 1, interval_color = "darkgrey", point_color = "verydark grey") +
  geom_point(data=intrafam, mapping = aes(x = Tribe, y = Diversity_island_natives), 
             shape = 19, size = 2,
             # color is hard coded here...so check/update if needed (Note! name in reverse order)
             colour = c("black")) +
  coord_flip() +
  theme_bw()+
  ggtitle("Observed tribe diversity vs null expectation") +
  xlab(NULL) +
  ylab(NULL)

tribe_null_expectation
```

```{r}
ggsave("../outputs/2_Diversity_patterns/Intrafamily_null_simulations.png", tribe_null_expectation, width = 150, height = 200, units = "mm")

ggsave("../outputs/2_Diversity_patterns/Intrafamily_null_simulations.svg", tribe_null_expectation, width = 150, height = 200, units = "mm")
```
